requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
gtts>=2.5.0
playsound>=1.3.0
requests>=2.31.0
pyaudio>=0.2.14
streamlit>=1.28.0
streamlit-chat>=0.1.1
importlib-metadata>=6.0.0,<7.0.0

